feat: repositoryに処理を分離
