+++ src/component/Panel/PostPanel.tsx
@@ -1,6 +1,5 @@
-import { db } from "@config/firebase";
+import { useTodo } from "@repo/todo/useTodo";
 import cc from "classcat";
-import { addDoc, collection } from "firebase/firestore";
 import React, { useState } from "react";
 
 import { PostButton } from "./PostButton";
@@ -17,47 +16,13 @@ type PostPanelPropsType = {
   value?: string;
 };
 
-type PostDataType = {
-  expireDate: Date;
-  name: string;
-  isDeleted: boolean;
-  isFinished: boolean;
-};
-
-// TODO: カスタムフックに分離するか検討中
-const postTaskData = async (expire: string, value: string) => {
-  const getExpireDate = (expire: string) => {
-    const now = new Date();
-    const tomorrow = new Date(now);
-    tomorrow.setDate(now.getDate() + 1);
-    const upcoming = new Date(now);
-    upcoming.setDate(now.getDate() + 9999); // TODO: 仮で9999日後に設定
-    switch (expire) {
-      case "today":
-        return now;
-      case "tomorrow":
-        return tomorrow;
-      case "upcoming":
-        return upcoming;
-      default:
-        return now;
-    }
-  };
-  const data: PostDataType = {
-    expireDate: getExpireDate(expire),
-    name: value ?? "",
-    isDeleted: false,
-    isFinished: false,
-  };
-  await addDoc(collection(db, "Todos"), data);
-};
-
 export const PostPanel: React.VFC<PostPanelPropsType> = ({
   onChange,
   onSubmit, // TODO: テキストボックスに入力された値を返すイベント（不要であれば削除します）
   value,
 }: PostPanelPropsType) => {
   const [textboxValue, setTextboxValue] = useState<string>(value ?? "");
+  const { onCreate } = useTodo();
 
   // TODO: テキストボックスに入力された値を返すイベント（不要の場合削除）
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -66,12 +31,12 @@ export const PostPanel: React.VFC<PostPanelPropsType> = ({
   };
 
   const handleSubmit = async (typeString: string) => {
+    await onCreate(typeString, textboxValue);
     onSubmit &&
       onSubmit({
         type: typeString || "",
         value: textboxValue || "",
       });
-    postTaskData(typeString, textboxValue);
     setTextboxValue("");
   };
   return (
