+++ src/repository/todo/todoCollection.ts
@@ -1,0 +1,1 @@
+export const todoCollection = (userId: string) => `user/${userId}/todos/`;
